hello

+++ ft_strlen.c
@@ -6,15 +6,15 @@
 /*   By: ecaruso <ecaruso@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/01/17 18:39:34 by ecaruso           #+#    #+#             */
-/*   Updated: 2023/01/19 16:05:41 by ecaruso          ###   ########.fr       */
+/*   Updated: 2023/01/21 20:13:20 by ecaruso          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int	ft_strlen(char *str)
+size_t	ft_strlen(char *str)
 {
-	int	a;
+	size_t	a;
 
 	a = 0;
 	while (str[a] != '\0')
